fix error

+++ manual_insert_once_crawling_python/crawling_aladin_book_rank.py
@@ -65,8 +65,9 @@ class AladinBookCrawling(crawling.Crawling, ABC):
                                charset=super().DB_CHARSET())
         curs = conn.cursor()
 
-        sql = """delete from aladin_book_rank"""
-        curs.execute(sql)
+        if rank_number == 1:
+            sql = """delete from aladin_book_rank"""
+            curs.execute(sql)
 
 
         sql = """insert into aladin_book_rank (rank, title, url, author, publisher, date, image_url) values (%s, %s, %s, %s, %s, %s, %s)"""
