+++ manual_insert_once_crawling_python/crawling_interpark_book_rank.py
@@ -48,8 +48,9 @@ class InterparkBookCrawling(crawling.Crawling, ABC):
                                charset=super().DB_CHARSET())
         curs = conn.cursor()
 
-        sql = """delete from interpark_book_rank"""
-        curs.execute(sql)
+        if rank_number == 1:
+            sql = """delete from interpark_book_rank"""
+            curs.execute(sql)
 
         sql = """insert into interpark_book_rank (rank, title, url, author, publisher, image_url) values (%s, %s, %s, %s, %s, %s)"""
         curs.execute(sql, (rank_number, book_title, book_info_url, book_author, book_publisher, image_url))
